fix: 倍速/rate 只接受整数档位——Float 传参挂死 TV 连接
根因：Kotlin Float.toString() 把 1f 渲染为 "1.0"、1.5f 为 "1.5"，
TV 固件 /rate 端点解析整数参数时收到带小数点的值直接挂死连接（不响应），
requestRaw 10s 超时后返回 null，被 request() 静默丢弃，UI 无报错。

curl 探针实证（TV-A 192.168.1.105:13520）：
  POST /rate?value=1 → 200 OK；value=3 → 200 OK
  POST /rate?value=1.0 → 超时挂死；value=1.5 → 超时挂死

改动：rate/volume 同模式，全部改为 Int 类型（Float→Int），
UI chips 从 [1, 1.5, 2, 3] 改为 [1, 2, 3]（TV 不支持半档）。
同步更新 docs/re/04 注明整数约束。

+++ shared/src/jvmCommonMain/kotlin/app/tvlink/ui/AppViewModel.kt
@@ -143,7 +143,7 @@ class AppViewModel : ViewModel() {
     var castDuration by mutableStateOf(0L)
     var castPosition by mutableStateOf(0L)
     var castVolume by mutableStateOf(0)
-    var castRate by mutableStateOf(1f)
+    var castRate by mutableStateOf(1)
     var castTitle by mutableStateOf("")
     var mediaServerUrl by mutableStateOf("")
     var castServerInfo by mutableStateOf<CastController.ServerInfo?>(null)
@@ -177,7 +177,7 @@ class AppViewModel : ViewModel() {
                     connectedIbSid = ""
                     mediaServerUrl = ""
                     castServerInfo = null
-                    castRate = 1f
+                    castRate = 1
                     cast?.disconnect()
                     cast = null
                     mediaServer.stop()
@@ -261,7 +261,7 @@ class AppViewModel : ViewModel() {
                     if (dur > 0) castDuration = dur
                     castPosition = pos
                     if (vol >= 0) castVolume = vol
-                    if (rate > 0f) castRate = rate
+                    if (rate > 0) castRate = rate
                 }
             }
             if (cc.connect()) {
@@ -409,7 +409,7 @@ class AppViewModel : ViewModel() {
     }
 
     /** 播放倍速（POST /rate）：乐观更新本地值，轮询随后校准（与音量同模式）。 */
-    fun castRateTo(r: Float) {
+    fun castRateTo(r: Int) {
         castRate = r
         viewModelScope.launch(Dispatchers.IO) { cast?.rate(r) }
     }
